Move TranslationForExistingLocalesValidator to LocaleBundle
