+++ src/Sylius/Bundle/LocaleBundle/Resources/translations/validators.en.yml
@@ -5,3 +5,6 @@ sylius:
             regex: Locale code can only be comprised of letters, numbers, dashes and underscores.
             locale: This value is not a valid locale code.
             unique: Locale code must be unique.
+    translation:
+        locale:
+            not_available: 'Please choose one of the available locales: %locales%'
