+++ src/Sylius/Bundle/LocaleBundle/spec/Validator/TranslationForExistingLocalesValidatorSpec.php
@@ -11,12 +11,12 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Bundle\CoreBundle\Validator\Constraints;
+namespace spec\Sylius\Bundle\LocaleBundle\Validator;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Sylius\Bundle\CoreBundle\Validator\Constraints\TranslationForExistingLocales;
+use Sylius\Bundle\LocaleBundle\Validator\TranslationForExistingLocales;
 use Sylius\Component\Locale\Model\LocaleInterface;
 use Sylius\Component\Resource\Model\TranslatableInterface;
 use Sylius\Component\Resource\Model\TranslationInterface;
